fix(search): toggle aria-expanded with the presence of autocomplete suggestions
By default, Downshift toggles aria-expanded with the focus state of the combobox. As we have revised
the code so autocomplete suggestions will be displayed with the presence of any search text, this
default behavior of Downshift has to be overriden.

re #414 (some issues are resolved; but we still need to mock Google Maps API to test the toggle of
aria-expanded)